♻️ [Refactor] 기타 리팩토링

+++ search-service/src/main/java/com/traveler/search/domain/post/document/PostDocument.java
@@ -78,7 +78,7 @@ public class PostDocument extends BaseDocument {
     private Long commentCount = 0L;
 
     @Field(type = FieldType.Rank_Feature)
-    private Long popularityScore;
+    private Double popularityScore;
 
     // popularityScore(rank_feature)는 relevance boost 전용이라 정렬/집계를 지원하지 않으므로,
     // 정렬·집계용으로 동일한 값을 별도의 숫자 필드에 복제해서 저장한다.
