+++ web-api-service/src/main/java/com/traveler/web/domain/search/dto/response/RankingResponse.java
@@ -7,11 +7,11 @@ public final class RankingResponse {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
     }
 
-    public record PostRank(Long postId, Long memberId, String title, String thumbnail, Long score) {}
+    public record PostRank(Long postId, Long memberId, String title, String thumbnail, Double score) {}
 
-    public record CategoryRank(String category, Long score) {}
+    public record CategoryRank(String category, Double score) {}
 
-    public record RegionRank(String region, Long score) {}
+    public record RegionRank(String region, Double score) {}
 
     public record LiveDTO(List<PostRank> posts, List<CategoryRank> categories, List<RegionRank> regions) {}
 }
